requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
asyncpg>=0.28.0
pydantic>=2.4.0
python-jose>=3.3.0
openai>=1.1.0
anthropic>=0.5.0
mistralai>=1.7.1
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-timeout>=2.1.0
httpx>=0.24.1
ruff>=0.1.0
python-multipart>=0.0.6
pgvector>=0.2.3
PyJWT>=2.0.0
cryptography>=45.0.0
cffi>=1.14
pycparser>=2.22
python-dotenv>=1.0.0
